Push mountpoint creation down into the client layer.
Previously mountpoint creation would happen in the CLI itself, which had
two downsides: (1) we created useless directories when using the mock
client; and (2) the mountpoint would have to be somewhere the user could
write to -- i.e. `/tmp`.

This commit pushes directory creation down into the client
implementations, which involves a few changes to the internal Client
trait in the process.

At the same time, we now create mountpoints in `/run` rather than `/tmp`
so that unmounted boot environments don't leave directories around on
reboot. This is a break from what `bectl` does, but seems much more in
line with Linux conventions.

We generate two-character directory names, which is a bit shorter than
you might expect, but it makes the `beadm list` output look very clean.