Update all client methods to take a boot environment root parameter.
This is part of an effort to make the D-Bus service capable of operating
on boot environment roots other than the active one.

To minimize the scope of the changes, this commit only updates the
client method signatures and the D-Bus method calls, the underlying
client implementations do not yet use the new parameter.

The init() method doesn't get a new parameter because it doesn't operate
on a boot environment.