Add an experimental 'beadm load' command for rebooting via kexec(8).
This commit adds direct integration with `kexec(8)`, which allows
rebooting the kernel without re-initialising firmware.

It largely follows the documented behaviour of ZFSBootMenu for detecting
matching kernal and initramfs pairs.

Unfortunately this approach relies heavily on the hardware in use to
actually reboot successfully, so it remains experimental for now.